Fixed bugs
- Fixed hamburger bug
- Sectioned header off as component to keep the metadata AND keep useState
